+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+* Added `shawl add` for quickly creating a Shawl-wrapped service.
+* Moved existing CLI functionality under `shawl run`.
+
 ## v0.2.0 (2019-09-22)
 
 * Send ctrl-C to child process first instead of always forcibly killing it.